api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "asyncpg>=0.30.0",
    "black>=25.11.0",
    "faker>=38.2.0",
    "fastapi[all]>=0.121.3",
    "flake8>=7.3.0",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "notebook>=7.5.0",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-socketio>=5.15.0",
    "rapidfuzz>=3.14.3",
    "sqlalchemy>=2.0.44",
    "stripe>=14.0.1",
    "structlog>=25.5.0",
    "websockets>=15.0.1",
    "openai>=2.8.1",
    "qrcode>=8.0",
    "pillow>=11.0.0",
]
